feat: Add side section
