+++ src/app/posts/[post_id]/page.style.ts
@@ -56,31 +56,6 @@ export const tags = styled.ul`
   }
 `;
 
-export const actionWrap = styled.ul`
-  ${({ theme }) => theme.MIXINS.flexBox('row', 'center', 'flex-start')};
-
-  li {
-    margin-left: 8px;
-    width: 32px;
-    height: 32px;
-    border-radius: 18px;
-    cursor: pointer;
-    background-color: ${({ theme }) => theme.colors.primary};
-    opacity: 0.5;
-    ${({ theme }) => theme.MIXINS.flexBox('row', 'center', 'center')};
-
-    &:hover {
-      opacity: 1;
-      transition: .3s;
-      transform: translateY(-4px);
-    }
-
-    svg {
-
-    }
-  }
-`;
-
 export const sideWrap = styled.div`
   left: 100%;
   position: absolute;
@@ -104,13 +79,5 @@ export const scrollspyWrap = styled.ul`
       color: ${({ theme }) => theme.colors.text['222222']};
       transition: .3s;
     }
-  }
-`;
-
-export const listWrap = styled.ul`
-  ${({ theme }) => theme.MIXINS.flexBox('column', 'flex-start', 'flex-start')};
-
-  ${({ theme }) => theme.media.pc} {
-    display: none;
   }
 `;
